Extract nikolaev_scq into separate class.
